@@ -1,33 +1,48 @@
-import {Component, EventEmitter, Input, OnChanges, Output} from '@angular/core';
+import {Component, computed, input, output} from '@angular/core';
 import {ComplexSearchResult} from '../../../shared/model/complex-results/complex-search.model';
 import {Interactor} from '../../../shared/model/complex-results/interactor.model';
 import {Element} from '../../../shared/model/complex-results/element.model';
 import {ComplexComponent} from '../../../shared/model/complex-results/complex-component.model';
+import * as tf from '@tensorflow/tfjs';
+import {groupByPropertyToArray} from '../../../complex-portal-utils';
 
 @Component({
   selector: 'cp-table-structure',
   templateUrl: './table-structure.component.html',
   styleUrls: ['./table-structure.component.css']
 })
-export class TableStructureComponent implements OnChanges {
-  @Input() complexSearch: ComplexSearchResult;
-  @Input() interactors: Interactor[];
-  @Input() interactorsSorting: string;
-  @Input() organismIconDisplay: boolean;
-  @Input() interactorTypeDisplay: boolean;
-  @Input() IDDisplay: boolean;
-  @Input() canAddComplexesToBasket: boolean;
-  @Input() canRemoveComplexesFromBasket: boolean;
-  @Output() onComplexRemovedFromBasket: EventEmitter<string> = new EventEmitter<string>();
+export class TableStructureComponent {
+  complexSearch = input<ComplexSearchResult>();
+  interactors = input<Interactor[]>();
+  interactorsSorting = input<string>();
+  organismIconDisplay = input<boolean>();
+  interactorTypeDisplay = input<boolean>();
+  idDisplay = input<boolean>();
+  canAddComplexesToBasket = input<boolean>();
+  canRemoveComplexesFromBasket = input<boolean>();
+  onComplexRemovedFromBasket = output<string>();
 
-  sortedComplexes: Element[] = [];
-
-  ngOnChanges(): void {
-    this.sortedComplexes = this.classifyComplexesSimilaritiesV2(this.complexSearch.elements);
-  }
+  sortedComplexes = computed(() => this.sortComplexBySimilarityClustering(this.complexSearch().elements));
 
   private getComponentAsComplex(component: ComplexComponent): Element | undefined {
-    return this.complexSearch.elements.find(interactor => interactor.complexAC === component.identifier);
+    return this.complexSearch().elements.find(interactor => interactor.complexAC === component.identifier);
+  }
+
+  private calculateSimilarity(complex1: Element, complex2: Element) {
+    if (complex1 === complex2) {
+      return 1;
+    }
+
+    const [components1, components2] = [complex1, complex2].map(this.getComponents.bind(this));
+    // @ts-ignore
+    return components1.intersection(components2).size / components1.union(components2).size;
+  }
+
+  private getComponents(complex: Element): Set<string> {
+    if (!complex.componentAcs) {
+      complex.componentAcs = new Set<string>(this.getAllComponents(complex).map(component => component.identifier));
+    }
+    return complex.componentAcs;
   }
 
   private getAllComponents(complex?: Element, components: ComplexComponent[] = []): ComplexComponent[] {
@@ -46,57 +61,58 @@ export class TableStructureComponent implements OnChanges {
     return components;
   }
 
-  private calculateSimilarity(complex1: Element, complex2: Element) {
-    if (complex1 === complex2) {
-      return new Set(this.getAllComponents(complex1)).size;
-    }
+  sortComplexBySimilarityClustering(complexesList: Element[]) {
+    // Group by predicted to cluster only inside the different groups, and place predicted after curated
+    const groups = groupByPropertyToArray(
+      complexesList,
+      'predictedComplex',
+      (a, b) => (a as unknown as number) - (b as unknown as number) // Makes false go before true
+    );
 
-    const [components1, components2] =
-      [complex1, complex2]
-        .map(complex => new Set(this.getAllComponents(complex).map(c => c.identifier)));
-
-    return [...components1.values()].reduce((s, c1) => components2.has(c1) ? s + 1 : s, 0);
+    return groups.flatMap(group => {
+      // Calculate Similarity Matrix
+      const sm: number[][] = new Array(group.length).fill(null).map(r => new Array(group.length).fill(null));
+      group.forEach((complex, i) => group.forEach((comparedComplex, j) => {
+        if (i >= j) { // Avoid useless calculations
+          sm[i][j] = this.calculateSimilarity(complex, comparedComplex);
+          sm[j][i] = sm[i][j]; // Matrix is symmetric
+        }
+      }));
+      return this.getSortedIndexFromChainedSimilarity(tf.tensor2d(sm)).map(i => group[i]);
+    });
   }
 
-  classifyComplexesSimilaritiesV2(complexesList: Element[]) {
-    const comparedComplexes: [Element, Element, number][] = [];
-    for (const complex of complexesList) {
-      for (const comparedComplex of complexesList) {
-        // for unique comparison
-        if (complex.complexAC >= comparedComplex.complexAC) {
-          comparedComplexes.push([complex, comparedComplex, this.calculateSimilarity(complex, comparedComplex)]);
-        }
-      }
-    }
-    comparedComplexes.sort((a, b) => b[2] - a[2]); // sorting by similarityScore
-    const complexesOrderedSet = this.uniqueComplexesListOrderedBySimilarity(comparedComplexes);
-    // to be used in the table as a 1D array
-    return Array.from(complexesOrderedSet);
-  }
+  /**
+   * Uses a similarity matrix to order elements based on how similar one element is to the previous element on the list.<br>
+   * Inspired by https://stackoverflow.com/a/64338609
+   *
+   * @param sm A similarity matrix to be sorted
+   * @returns indexes of elements ordered according to their similarity
+   */
+  getSortedIndexFromChainedSimilarity(sm: tf.Tensor2D): number[] {
+    const idx: number[] = [];
+    let currentIdx = tf.argMax(sm.sum(0)).arraySync() as number;  // Start with complex with most similarities
+    idx.push(currentIdx);
 
-  uniqueComplexesListOrderedBySimilarity(complexesListSimilarities: [Element, Element, number][]) {
-    const complexesOrderedSet = new Set<Element>();
-    for (let i = 0; i < complexesListSimilarities.length; i++) {
-      const [complex1, complex2, similarityScore] = complexesListSimilarities[i];
-      if (similarityScore !== 0) {
-        complexesOrderedSet.add(complex1);
-        complexesOrderedSet.add(complex2);
-        for (let j = i + 1; j < complexesListSimilarities.length; j++) {
-          const [complex3, complex4, similarityScore2nd] = complexesListSimilarities[j];
-          if (complex1 === complex3 && similarityScore2nd !== 0) {
-            complexesOrderedSet.add(complex4);
-          }
-        }
-        if (complexesOrderedSet.size === this.complexSearch.elements.length) {
-          // All complexes have been added, we can return and stop the loops
-          return complexesOrderedSet;
-        }
+    for (let i = 1; i < sm.shape[0]; i++) {
+      // Mask already selected indices
+      const buffer = sm.bufferSync();
+      for (let x = 0; x < sm.shape[0]; x++) {
+        buffer.set(-Infinity, x, currentIdx);
       }
-      if (complexesOrderedSet.size === this.complexSearch.elements.length) {
-        // All complexes have been added, we can return and stop the loops
-        return complexesOrderedSet;
+      sm = buffer.toTensor();
+
+      // Find the closest complex to last selected
+      const sm_i = sm.gather([currentIdx]).arraySync()[0] as number[];
+      currentIdx = tf.argMax(sm_i).arraySync() as number;
+
+      // If no similar entities (cluster ends), take the next complex with most similarities
+      if (sm_i[currentIdx] === 0) {
+        currentIdx = tf.argMax(sm.sum(0)).arraySync() as number;
       }
+
+      idx.push(currentIdx);
     }
-    return complexesOrderedSet;
+    return idx;
   }
 }
