@@ -1,8 +1,8 @@
-import {HttpClient, HttpErrorResponse} from '@angular/common/http';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {environment} from '../../../../environments/environment';
 import {catchError, map} from 'rxjs/operators';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {throwError} from 'rxjs/internal/observable/throwError';
 
 const baseURL = environment.ols_base_url;
@@ -32,6 +32,16 @@ export class OlsService {
    */
   getEfoName(id: string) {
     return this.http.get(baseURL + '/efo/terms?iri=http://www.ebi.ac.uk/efo/' + id.replace(':', '_')).pipe(
+      catchError(this.handleError));
+  }
+
+  /**
+   * Get a name of mondo xref
+   * @param id
+   * @returns {Observable<R>}
+   */
+  getOboName(id: string) {
+    return this.http.get(baseURL + '/efo/terms?iri=http://purl.obolibrary.org/obo/' + id.replace(':', '_')).pipe(
       catchError(this.handleError));
   }
 
