@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, Input, OnInit, input } from '@angular/core';
 import {CrossReference} from '../../shared/model/complex-details/cross-reference.model';
 
 @Component({
@@ -11,7 +11,7 @@ export class ComplexReferencesComponent implements OnInit {
   private _pubmedXRefs: CrossReference[];
   private _externalResources: CrossReference[];
   private _synonyms: string[];
-  private _systematicName: string[];
+  private _systematicName: string;
 
   constructor() {
   }
@@ -23,7 +23,7 @@ export class ComplexReferencesComponent implements OnInit {
   private findXRefs() {
     for (let i = 0; i < this._crossReferences.length; i++) {
       const crossRef = this._crossReferences[i];
-      const database = this._crossReferences[i].database;
+      const database = this._crossReferences[i].database.toLowerCase();
 
       if (database === 'pubmed') {
         if (this._pubmedXRefs === undefined) {
@@ -31,7 +31,7 @@ export class ComplexReferencesComponent implements OnInit {
         }
         this._pubmedXRefs.push(crossRef);
       }
-      if (database === 'signor' || database === 'matrixdb') {
+      if (database === 'signor' || database === 'matrixdb' || database === 'humap') {
         if (this._externalResources === undefined) {
           this._externalResources = [];
         }
@@ -74,12 +74,12 @@ export class ComplexReferencesComponent implements OnInit {
     this._synonyms = value;
   }
 
-  get systematicName(): string[] {
+  get systematicName(): string {
     return this._systematicName;
   }
 
   @Input()
-  set systematicName(value: string[]) {
+  set systematicName(value: string) {
     this._systematicName = value;
   }
 }
