@@ -1,5 +1,5 @@
 import {AfterViewInit, Component, OnInit} from '@angular/core';
-import {ActivatedRoute, NavigationExtras, Router} from '@angular/router';
+import {ActivatedRoute, NavigationExtras, Params, Router} from '@angular/router';
 import {ComplexSearchResult} from '../shared/model/complex-results/complex-search.model';
 import {ComplexPortalService} from '../shared/service/complex-portal.service';
 import {ProgressBarComponent} from '../../shared/loading-indicators/progress-bar/progress-bar.component';
@@ -7,10 +7,8 @@ import {Title} from '@angular/platform-browser';
 import {AnalyticsService} from '../../shared/google-analytics/service/analytics.service';
 import {Interactor} from '../shared/model/complex-results/interactor.model';
 import {NotificationService} from '../../shared/notification/service/notification.service';
-import {
-  COMPLEX_NAVIGATOR_VIEW,
-  LIST_VIEW
-} from './complex-navigator/complex-list-display-buttons/complex-list-display-buttons.component';
+import {SearchDisplay} from './complex-navigator/complex-list-display-buttons/complex-list-display-buttons.component';
+import {switchMap, tap} from 'rxjs/operators';
 
 @Component({
   selector: 'cp-complex-results',
@@ -18,15 +16,21 @@ import {
   styleUrls: ['./complex-results.component.css'],
 })
 export class ComplexResultsComponent implements OnInit, AfterViewInit {
-  private _query: string;
-  private _complexSearch: ComplexSearchResult;
-  private _spicesFilter: string[];
-  private _bioRoleFilter: string[];
-  private _interactorTypeFilter: string[];
-  private _allInteractorsInComplexSearch: Interactor[] = [];
-  DisplayType: string;
-  toast;
+  query: string;
+  complexSearch: ComplexSearchResult;
 
+  allInteractorsInComplexSearch: Interactor[] = [];
+  _displayType: SearchDisplay;
+
+  filters = {
+    species: [],
+    bioRole: [],
+    interactorType: [],
+    predicted: [],
+    confidenceScore: [],
+  };
+
+  private _toast;
   private _listPageSize = 15; // This is where we set the size of the pages for list view
   private _navigatorPageSize = 20; // This is where we set the size of the pages for navigator view
   private _listCurrentPage: number;
@@ -41,26 +45,24 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
 
   ngOnInit() {
     this.titleService.setTitle('Complex Portal - Results');
-    this._allInteractorsInComplexSearch = [];
-    this.route.fragment.subscribe(fragment => {
-      if (fragment === COMPLEX_NAVIGATOR_VIEW) {
-        this.DisplayType = COMPLEX_NAVIGATOR_VIEW;
-      } else if (fragment === LIST_VIEW) {
-        this.DisplayType = LIST_VIEW;
-      }
-      this.route
-        .queryParams
-        .subscribe(queryParams => {
-          this._query = queryParams['query'];
-          this._spicesFilter = queryParams['species'] ? queryParams['species'].split('+') : [];
-          this._bioRoleFilter = queryParams['bioRole'] ? queryParams['bioRole'].split('+') : [];
-          this._interactorTypeFilter = queryParams['interactorType'] ? queryParams['interactorType'].split('+') : [];
-          this.currentPageIndex = queryParams['page'] ? Number(queryParams['page']) : 1;
-          // TODO This is out for now, but CP-84 (JIRA )should fix that!!
-          // this.pageSize = queryParams['size'] ? Number(queryParams['size']) : 10;
-          this.requestComplexResults();
-          document.body.scrollTop = 0;
-        });
+    this.allInteractorsInComplexSearch = [];
+    this.route.fragment.pipe(
+      tap(f => {
+        if (f === SearchDisplay.navigator) {
+          this._displayType = SearchDisplay.navigator;
+        } else if (f === SearchDisplay.list) {
+          this._displayType = SearchDisplay.list;
+        }
+      }),
+      switchMap(_ => this.route.queryParams)
+    ).subscribe(queryParams => {
+      this.query = queryParams['query'];
+      Object.keys(this.filters).forEach(filter => this.filters[filter] = this.decodeURL(filter, queryParams));
+      this.currentPageIndex = queryParams['page'] ? Number(queryParams['page']) : 1;
+      // TODO This is out for now, but CP-84 (JIRA )should fix that!!
+      // this.pageSize = queryParams['size'] ? Number(queryParams['size']) : 10;
+      this.requestComplexResults();
+      document.body.scrollTop = 0;
     });
   }
 
@@ -69,16 +71,18 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
 
 
   private requestComplexResults() {
-    this.complexPortalService.findComplex(this.query, this.spicesFilter, this.bioRoleFilter,
-      this.interactorTypeFilter, this.currentPageIndex, this.pageSize).subscribe(complexSearch => {
+    this.complexPortalService.findComplex(this.query, this.filters.species, this.filters.bioRole,
+      this.filters.interactorType, this.filters.predicted, this.filters.confidenceScore,
+      this.currentPageIndex, this.pageSize).subscribe(complexSearch => {
       this.complexSearch = complexSearch;
-      this._allInteractorsInComplexSearch = [];
+      this.processSearchResults();
+      this.allInteractorsInComplexSearch = [];
       if (this.complexSearch.totalNumberOfResults !== 0) {
         this.lastPageIndex = Math.ceil(complexSearch.totalNumberOfResults / this.pageSize);
         for (let i = 0; i < complexSearch.elements.length; i++) {
           for (const component of complexSearch.elements[i].interactors) {
-            if (!this._allInteractorsInComplexSearch.some(interactor => interactor.identifier === component.identifier)) {
-              this._allInteractorsInComplexSearch.push(component);
+            if (!this.allInteractorsInComplexSearch.some(interactor => interactor.identifier === component.identifier)) {
+              this.allInteractorsInComplexSearch.push(component);
             }
           }
         }
@@ -92,20 +96,14 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
    */
   private reloadPage(): void {
     const queryParams: NavigationExtras = {};
-    queryParams['query'] = this._query;
+    queryParams['query'] = this.query;
     queryParams['page'] = this.currentPageIndex;
-    if (this._spicesFilter !== undefined && this._spicesFilter.length !== 0) {
-      queryParams['species'] = this.prepareFiltersForParams(this.spicesFilter);
-    }
-    if (this._bioRoleFilter !== undefined && this._bioRoleFilter.length !== 0) {
-      queryParams['bioRole'] = this.prepareFiltersForParams(this._bioRoleFilter);
-    }
-    if (this._interactorTypeFilter !== undefined && this._interactorTypeFilter.length !== 0) {
-      queryParams['interactorType'] = this.prepareFiltersForParams(this._interactorTypeFilter);
-    }
+
+    Object.keys(this.filters).forEach(filter => this.encodeURL(this.filters[filter], filter, queryParams));
+
     this.router.navigate([], {
       queryParams,
-      fragment: this.DisplayType
+      fragment: this.displayType
     });
     ProgressBarComponent.hide();
     // This is a test case event for GA, to monitor if users ever use more then one filter.
@@ -115,12 +113,19 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
     }
   }
 
-  private prepareFiltersForParams(filter: string[]): string {
-    return filter.toString().replace(/,/g, '+');
+  private encodeURL(filter: string[], filterName: string, params: Params): string {
+    if (filter !== undefined && filter.length !== 0) {
+      params[filterName] = filter.join().replace(/ /g, '_');
+    }
+    return params[filterName];
+  }
+
+  private decodeURL(filterName: string, params: Params): string[] {
+    return params[filterName] ? params[filterName].replace(/_/g, ' ').split(',') : [];
   }
 
   private getFilterCount(): number {
-    return this._spicesFilter.length + this._interactorTypeFilter.length + this._bioRoleFilter.length;
+    return Object.values(this.filters).reduce((total, f) => total + f.length, 0);
   }
 
   /**
@@ -133,41 +138,43 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
   }
 
   public onResetAllFilters(): void {
-    this.spicesFilter = [];
-    this.bioRoleFilter = [];
-    this.interactorTypeFilter = [];
+    Object.keys(this.filters).forEach(f => this.filters[f] = []);
     this.currentPageIndex = 1;
     this.reloadPage();
   }
 
-  public onSpicesFilterChanged(filter: string[]): void {
-    this.spicesFilter = filter;
+  public onSpeciesFilterChanged(filter: string[]): void {
+    this.filters.species = filter;
     this.currentPageIndex = 1;
     this.reloadPage();
   }
 
   public onBiologicalRoleFilterChanged(filter: string[]): void {
-    this.bioRoleFilter = filter;
+    this.filters.bioRole = filter;
     this.currentPageIndex = 1;
     this.reloadPage();
   }
 
   public onInteractorTypeFilterChanged(filter: string[]): void {
-    this.interactorTypeFilter = filter;
+    this.filters.interactorType = filter;
     this.currentPageIndex = 1;
     this.reloadPage();
   }
 
-  get query(): string {
-    return this._query;
+  public onPredictedFilterChanged(filter: string[]): void {
+    this.filters.predicted = filter;
+    this.currentPageIndex = 1;
+    this.reloadPage();
   }
 
-  set query(value) {
-    this._query = value;
+  public onConfidenceScoreFilterChanged(filter: string[]): void {
+    this.filters.confidenceScore = filter;
+    this.currentPageIndex = 1;
+    this.reloadPage();
   }
 
   get currentPageIndex(): number {
-    if (this.DisplayType === COMPLEX_NAVIGATOR_VIEW) {
+    if (this._displayType === SearchDisplay.navigator) {
       return this._navigatorCurrentPage;
     } else {
       return this._listCurrentPage;
@@ -175,24 +182,15 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
   }
 
   set currentPageIndex(value: number) {
-    if (this.DisplayType === COMPLEX_NAVIGATOR_VIEW) {
+    if (this._displayType === SearchDisplay.navigator) {
       this._navigatorCurrentPage = value;
     } else {
       this._listCurrentPage = value;
     }
   }
 
-  get complexSearch(): ComplexSearchResult {
-    return this._complexSearch;
-  }
-
-  set complexSearch(value: ComplexSearchResult) {
-    this._complexSearch = value;
-    this.processSearchResults();
-  }
-
   get lastPageIndex(): number {
-    if (this.DisplayType === COMPLEX_NAVIGATOR_VIEW) {
+    if (this._displayType === SearchDisplay.navigator) {
       return this._navigatorLastPageIndex;
     } else {
       return this._listLastPageIndex;
@@ -200,7 +198,7 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
   }
 
   set lastPageIndex(value: number) {
-    if (this.DisplayType === COMPLEX_NAVIGATOR_VIEW) {
+    if (this._displayType === SearchDisplay.navigator) {
       this._navigatorLastPageIndex = value;
     } else {
       this._listLastPageIndex = value;
@@ -208,66 +206,44 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
   }
 
   get pageSize(): number {
-    if (this.DisplayType === COMPLEX_NAVIGATOR_VIEW) {
+    if (this._displayType === SearchDisplay.navigator) {
       return this._navigatorPageSize;
     } else {
       return this._listPageSize;
     }
   }
 
-  get spicesFilter(): string[] {
-    return this._spicesFilter;
-  }
-
-  set spicesFilter(value: string[]) {
-    this._spicesFilter = value;
-  }
-
-  get bioRoleFilter(): string[] {
-    return this._bioRoleFilter;
-  }
-
-  set bioRoleFilter(value: string[]) {
-    this._bioRoleFilter = value;
-  }
-
-  get interactorTypeFilter(): string[] {
-    return this._interactorTypeFilter;
-  }
-
-  set interactorTypeFilter(value: string[]) {
-    this._interactorTypeFilter = value;
-  }
-
-  public get allInteractorsInComplexSearch(): Interactor[] {
-    return this._allInteractorsInComplexSearch;
-  }
-
-  onDisplayTypeChange(displayType: string) {
-    if (this.DisplayType !== displayType) {
-      this.DisplayType = displayType;
-      if (displayType === LIST_VIEW) {
+  set displayType(displayType: SearchDisplay) {
+    if (this._displayType !== displayType) {
+      this._displayType = displayType;
+      if (displayType === SearchDisplay.list) {
         this.setListView();
-      } else if (displayType === COMPLEX_NAVIGATOR_VIEW) {
+      } else if (displayType === SearchDisplay.navigator) {
         this.setComplexNavigatorView();
       }
     }
   }
 
+  get displayType(): SearchDisplay {
+    return this._displayType;
+  }
+
   isDisplayComplexNavigatorView(): boolean {
-    return this.DisplayType === COMPLEX_NAVIGATOR_VIEW;
+    return this._displayType === SearchDisplay.navigator;
   }
 
   private setListView() {
-    this.toast = this.notificationService.complexNavigatorAnnouncement();
+    this._toast = this.notificationService.complexNavigatorAnnouncement();
+    this._displayType = SearchDisplay.list;
     this.reloadPage();
   }
 
   private setComplexNavigatorView() {
-    if (!!this.toast) {
-      this.notificationService.closeAnnouncement(this.toast.toastId);
-      this.toast = null;
+    if (!!this._toast) {
+      this.notificationService.closeAnnouncement(this._toast.toastId);
+      this._toast = null;
     }
+    this._displayType = SearchDisplay.navigator;
     this.reloadPage();
   }
 
@@ -275,19 +251,15 @@ export class ComplexResultsComponent implements OnInit, AfterViewInit {
     // No filters and only one result, then we redirect to complex details page
     // This allows users to enable filters to see even one result without redirecting them out from the results page,
     // but we ensure redirection of a new search has only one result.
-    if (this.getFilterCount() === 0 && this._complexSearch.totalNumberOfResults === 1) {
-      const complexId = this._complexSearch.elements[0].complexAC;
+    if (this.getFilterCount() === 0 && this.complexSearch.totalNumberOfResults === 1) {
+      const complexId = this.complexSearch.elements[0].complexAC;
       if (!!complexId) {
-        // For some reason this is needed so the navigate call works
-        this.router.routeReuseStrategy.shouldReuseRoute = function () {
-          return false;
-        };
         this.router.navigate(['/complex', complexId]);
       }
-    } else if (!this.DisplayType) {
+    } else if (!this._displayType) {
       // Currently the list view is the default, as we are just launching the navigator view
       // Later on we can change the default view to be the list or navigator view based on number of results
-      if (this._complexSearch.totalNumberOfResults <= this._navigatorPageSize) {
+      if (this.complexSearch.totalNumberOfResults <= this._navigatorPageSize) {
         this.setComplexNavigatorView();
       } else {
         this.setListView();
